Use disconnectBody in easyrtc.disconnect
This should fix a few issues, such as the pre allocated socket getting
closed on disconnect, the websocket not being reset after the close,
the roomOccupant listener getting called with room name set to null,
and the disconection listener not getting called.

It also streamlines the disconnection event and avoids code duplication.