Always close the socket when disconnecting
By closing the socket on disconnect even if we've not established
a connection, it allows us to actually cancel a connection attempt
that hasn't succeeded yet. Otherwise we end up with callbacks being
called when webSocket is set to 0 and potentially opening a new
connection after the user tried to cancel the previous one.